requirements.txt
opencv-python
python-dotenv
ultralytics

--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio
numpy
